feat(markdown): Wholesale markdown adjustments
