+++ package.json
@@ -21,11 +21,14 @@
   "devDependencies": {
     "@11ty/eleventy": "^0.11.0",
     "@11ty/eleventy-plugin-rss": "^1.0.7",
-    "markdown-it": "^10.0.0",
-    "markdown-it-anchor": "^5.3.0",
-    "node-sass": "^4.13.1",
+    "@toycode/markdown-it-class": "^1.2.4",
     "eleventy-plugin-reading-time": "0.0.1",
-    "html-minifier": "^4.0.0"
+    "highlight.js": "^10.1.1",
+    "html-minifier": "^4.0.0",
+    "markdown-it": "^10.0.0",
+    "markdown-it-abbr": "^1.0.4",
+    "markdown-it-anchor": "^5.3.0",
+    "node-sass": "^4.13.1"
   },
   "dependencies": {}
 }
