+++ src/_includes/layouts/post.njk
@@ -16,6 +16,8 @@
 
 {% from "components/hero.html" import hero %}
 
+<link href="https://fonts.googleapis.com/css2?family=Fira+Mono:wght@400;500&display=swap" rel="stylesheet">
+
 {% set children = "" %}
 {% set children %}
   <p class="post-description"><strong>{{ page.date | formatDate }}</strong>{% if description %} {{ description | safe }}{% endif %}</p>
