+++ src/_includes/scss/_reset.scss
@@ -10,11 +10,6 @@ dt {
   margin: 0;
 }
 
-abbr {
-  text-decoration: none;
-  border-bottom: 0;
-}
-
 strong {
   font-weight: 500;
 }
