+++ src/_includes/scss/_type.scss
@@ -40,53 +40,6 @@
   scroll-margin-top: 1.25em;
 }
 
-%blockquote {
-  font-family: 'rsa-heading', serif;
-  font-weight: 400;
-  font-style: italic;
-  font-size: clamp(1.5rem, calc(1rem + 1vw), 2.5rem);
-  margin: 2em 0;
-  line-height: 1.3;
-  max-width: 24em;
-
-  cite {
-    @extend %lead;
-
-    font-size: 1.25rem;
-    font-family: 'rsa-body', sans-serif;
-    font-style: normal;
-    margin-top: 0.25em;
-    display: block;
-
-    a {
-      @extend %lead-a;
-    }
-  }
-
-  &::before {
-    content: '“';
-    color: var(--c-bA);
-    display: block;
-    font-weight: 600;
-    font-style: normal;
-    width: 0.5em;
-    left: 0;
-    top: 0.05em;
-    font-size: 1.5em;
-    position: relative;
-    line-height: 0.5;
-  }
-
-  > p {
-    position: relative;
-    margin-bottom: 0.5em;
-
-    &:last-child {
-      margin-bottom: 0;
-    }
-  }
-}
-
 %lead {
   @include sc;
