+++ src/_includes/scss/_typography.scss
@@ -12,7 +12,7 @@ p,
 }
 
 del {
-  text-decoration-color: var(--c-bA);
+  text-decoration-color: var(--c-bC-66);
   text-decoration-thickness: 2px;
 }
 
@@ -33,7 +33,106 @@ del {
 }
 
 .blockquote {
-  @extend %blockquote;
+  font-family: 'rsa-heading', serif;
+  font-weight: 400;
+  font-style: italic;
+  font-size: clamp(1.5rem, calc(1rem + 1vw), 2.5rem);
+  margin: 2em 0;
+  line-height: 1.3;
+  max-width: 24em;
+
+  cite {
+    @extend %lead;
+
+    font-size: 1.25rem;
+    font-family: 'rsa-body', sans-serif;
+    font-style: normal;
+    margin-top: 0.25em;
+    display: block;
+
+    &::before {
+      content: '— ';
+    }
+
+    a {
+      @extend %lead-a;
+    }
+  }
+
+  &::before {
+    content: '“';
+    color: var(--c-bA);
+    display: block;
+    font-weight: 600;
+    font-style: normal;
+    width: 0.5em;
+    left: 0;
+    top: 0.05em;
+    font-size: 1.5em;
+    position: relative;
+    line-height: 0.5;
+  }
+
+  > p {
+    position: relative;
+    margin-bottom: 0.5em;
+
+    &:last-child {
+      margin-bottom: 0;
+    }
+  }
+}
+
+abbr {
+  border-bottom: 0;
+  text-decoration: none;
+}
+
+code {
+  color: var(--c-bA);
+  font-weight: 400;
+  font-size: 0.85em;
+  font-family: 'Fira Mono', monospace;
+
+  a & {
+    color: inherit;
+    font-weight: 500;
+  }
+}
+
+.code-block {
+  background: var(--c-hB);
+  clear: both;
+  padding: 1.8em 2em 2em;
+  font-size: 1rem;
+  white-space: pre-wrap;
+  margin: 2rem 0;
+  max-width: 48em;
+
+  &[data-lang] {
+    position: relative;
+
+    &::before {
+      @extend .sc;
+
+      content: attr(data-lang);
+      color: var(--c-bC-66);
+      font-family: 'rsa-body', sans-serif;
+      position: absolute;
+      top: 0;
+      right: 0.5em;
+      font-weight: 500;
+      font-size: 0.75rem;
+    }
+  }
+
+  code {
+    line-height: 1.85;
+  }
+}
+
+.hljs-comment {
+  opacity: 0.66;
 }
 
 .lead {
